% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{light_cool_colors}
\alias{light_cool_colors}
\title{light_cool_colors}
\usage{
light_cool_colors(...)
}
\arguments{
\item{...}{Arguments passed to the function returned by `colorRampPalette`}
}
\value{
A light color palette function
}
\description{
A light color palette.
}
\details{
A simple light color palette gradient from dark blue to light blue 
intended for a heatmap with a white or light color background. 
 
 Equivalent to `colorRampPalette(c("#1133AA", "#CCFFFF"))(...)`.
}
\examples{
light_colors <- light_cool_colors(144)
plot(1:144, 1:144, col = light_colors, pch = 19)
}
