% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{light_heat_colors}
\alias{light_heat_colors}
\title{light_heat_colors}
\usage{
light_heat_colors(...)
}
\arguments{
\item{...}{Arguments passed to the function returned by `colorRampPalette`}
}
\value{
A light color palette function
}
\description{
A light color palette.
}
\details{
A simple light color palette gradient from red to orange to gold to
 yellow intended for a heatmap with a white or light color background. 
 
 Equivalent to `colorRampPalette(c("red", "darkorange2", "darkgoldenrod1", "gold1", "yellow2"))(...)`.
}
\examples{
light_colors <- light_heat_colors(144)
plot(1:144, 1:144, col = light_colors, pch = 19)
}
