% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{relxy}
\alias{relxy}
\alias{relx}
\alias{rely}
\title{relxy}
\usage{
relx(r)

rely(r)
}
\arguments{
\item{r}{Radius of point data relative to X or Y range of the plot, a value between 0 and 1.}
}
\value{
A class structure for input to `GlowMapper$map`
}
\description{
Helper functions for specifying the `radius` parameter in `GlowMapper$map` and similar.
}
\details{
Helper functions for specifying the `radius` parameter relative to the range of the plot as a proportion. 
I.e., a value of 0.02 corresponds to 2% of the plot.
}
\examples{

gm <- GlowMapper4$new(xdim=240, ydim=240, blend_mode = "additive", nthreads=1)
gm$map(x=cliff_points$x, y=cliff_points$y, radius=relx(0.02), color=cliff_points$color)
pd <- gm$output_dataframe(saturation = 1)

ggplot() +
  geom_raster(data = pd, aes(x = x, y = y, fill = rgb(r,g,b,a)), show.legend=FALSE) +
  coord_fixed(gm$aspect(), xlim = gm$xlim(), ylim = gm$ylim()) + 
  scale_fill_identity() + 
  theme_night()
}
