% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_help_files.R
\name{theme_night}
\alias{theme_night}
\title{theme_night}
\usage{
theme_night(bgcolor = "black", base_size = 14, base_family = "")
}
\arguments{
\item{bgcolor}{Background color, default black. Generally you want to match the background with the lowest color value on a color scale.}

\item{base_size}{Base default font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme.
}
\description{
A dark ggplot2 theme with a default black background intended to be used with the glow package.
}
\details{
The theme is heavily modified from the minimal ggplot theme. It is intended to be use with dark background colors and should not be used with white or light backgrounds.
}
\examples{
ggplot(mtcars, aes(x = mpg, y = wt)) + 
  geom_point(color = "white") +
  theme_night(bgcolor = "black")
}
